requirements.txt
allure-pytest==2.14.1
allure-python-commons==2.14.1
attrs==25.1.0
certifi==2025.1.31
coverage==7.6.10
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
pytest-cov==6.0.0
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
webdriver-manager>=3.8.0
requests~=2.32.3